[OVERHAUL] iterator now uses the classes built to refactor it
